3-3-2023 -Mason Peters
Fixed "startingBudget" variable and it should update every time the input is changed

Added actual variables for the settings, that way it can be accessed in the game and determine anything it needs to determine.

Added the presets for "difficulty", that way the startingBudget value can be preset depending on what the player chooses for a difficulty. Also works with custom.

Polished up the look for event1 & 2

Added a new image